Use parens to surround init arguments
This means the init block no longer resembles closure syntax, but it
makes it possible to use empty argument lists.

It's would possible to support empty arguments while preserving the
current syntax, but that seems more complicated due to the kinds of
method provided by the parser. Right now, it's not really worth the
effort, I think.