Rename World to Entities
Entities seems the more fitting name, since managing entities is what it
does (and all it does). I originally imagined that more things would be
added (like support for systems, events), but my current thinking is
that this functionality belongs with other data types. We'll see, I can
always change it back.

I'm leaving the macro name as is for now. Right now, it's not very clear
to me what responsibilities it'll eventually have.