Remove implicit Clone requirement
Due to a call to HashMap::find_copy, components still needed to
implement Clone. This has been fixed by this commit.

The side effect of this fix is that Entities::export consumes self. If
the user doesn't want that, they can just explicitely clone Entities
before the export. Since before, each component was cloned during the
export, this isn't much different and makes the cost of cloning more
explicit.

Long-term, there should be a better solution for exporting entities that
doesn't require all those allocations, but for now, the current behavior
is good enough.